DI-456 Resolved circular dependencies in change event (#334)
* Resolved circular dependencies in change event

* Further amendments to resolve dependency issues

* Applying changes from commit checks
