+++ test/integration/steps/utilities/change_request.py
@@ -1,0 +1,34 @@
+from typing import Any, Dict
+
+
+def change_request() -> Dict[str, Any]:
+    return {
+        "change_payload": {
+            "reference": "EDFA07-16",
+            "system": "DoS Integration",
+            "message": "DoS Integration CR. correlation-id: EDFA07-16",
+            "service_id": "37652",
+            "changes": {
+                "website": None,
+                "phone": None,
+                "public_name": "My Test Pharmacy 21",
+                "address": {
+                    "address_lines": ["85 Peachfield Road", "CHAPEL ROW", "South Godshire"],
+                    "post_code": "RG7 1DB",
+                },
+                "opening_days": {
+                    "Monday": [],
+                    "Tuesday": [],
+                    "Wednesday": [],
+                    "Thursday": [],
+                    "Friday": [],
+                    "Saturday": [],
+                    "Sunday": [],
+                },
+            },
+        },
+        "correlation_id": "c1",
+        "message_received": 1643306908893,
+        "dynamo_record_id": "d8842511670361f8db0f52d5ab86e78c",
+        "ods_code": "FA007",
+    }
